add get public garden by coordinate
- add service logic to fetch from repository and convert to response
- add jpql query for fetching public garden by coordinate(start, end)
- modify longitude, latitude data type from int to double

+++ everyone-garden/src/main/java/com/everyonegarden/garden/Garden.java
@@ -21,8 +21,8 @@ public class Garden {
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    private Integer latitude;
-    private Integer longitude;
+    private double latitude;
+    private double longitude;
     private String address;
 
     @Column(nullable = false) @NotBlank
